# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: NexBank Digital Banking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List user accounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Accounts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /accounts/stats:
    get:
      operationId: getAccountStats
      tags: [accounts]
      summary: Get total balance and account count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountStatsResponse"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account nickname
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Account updated
  /accounts/{id}/freeze:
    post:
      operationId: freezeAccount
      tags: [accounts]
      summary: Freeze account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account frozen
  /accounts/{id}/unfreeze:
    post:
      operationId: unfreezeAccount
      tags: [accounts]
      summary: Unfreeze account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account unfrozen
  /accounts/{id}/transactions:
    get:
      operationId: getAccountTransactions
      tags: [accounts]
      summary: Get transactions for an account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPageResponse"

  # ── TRANSACTIONS ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountNumber:
          type: string
        accountType:
          type: string
          enum: [CHECKING, SAVINGS, MONEY_MARKET, FIXED_DEPOSIT]
        balance:
          type: string
        currency:
          type: string
        status:
          type: string
          enum: [ACTIVE, FROZEN, CLOSED]
        nickname:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        routingNumber:
          type: string
        interestRate:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, accountNumber, accountType, balance, currency, status]
    AccountInput:
      type: object
      properties:
        accountType:
          type: string
          enum: [CHECKING, SAVINGS, MONEY_MARKET, FIXED_DEPOSIT]
        currency:
          type: string
        nickname:
          type: ["string", "null"]
    AccountUpdate:
      type: object
      properties:
        nickname:
          type: ["string", "null"]
    AccountResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Account"
    AccountListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Account"
    AccountStatsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            totalBalance:
              type: string
            accountCount:
              type: integer

    # ── Transaction Schemas ───────────────────────────────────────────
    Transaction:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        userId:
          type: integer
        refNumber:
          type: string
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, TRANSFER_IN, TRANSFER_OUT, FEE, INTEREST, REVERSAL]
        amount:
          type: string
        balanceBefore:
          type: string
        balanceAfter:
          type: string
        currency:
          type: string
        description:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, REVERSED]
        counterpartName:
          type: ["string", "null"]
        counterpartBank:
          type: ["string", "null"]
        counterpartAccountNumber:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, refNumber, type, amount, status]
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...
    TransactionPageResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Transaction"
            pagination:
              $ref: "#/components/schemas/Pagination"
...