# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NexBank Digital Banking API
 * OpenAPI spec version: 1.0.0
 */
import type { AccountInputAccountType } from './accountInputAccountType';

export interface AccountInput {
  accountType?: AccountInputAccountType;
  currency?: string;
  /** @nullable */
  nickname?: string | null;
}